Reference

Contents

Reference

Contents

Index

Index

FloatsForML.CODE_TYPES

FloatsForML.FLOAT_TYPES

FloatsForML.CODE

FloatsForML.FLOAT

FloatsForML.MLFloats

FloatsForML.index_to_offset

FloatsForML.offset_to_index

FloatsForML.typeforcode

FloatsForML.typeforfloat

FloatsForML.CODE — Type

CODE

The built-in Unsigned Integer types available for encodings.

UInt8 for bitwidths <= 8

UInt16 for bitwidths > 8

see also FLOAT

FloatsForML.CODE_TYPES — Constant

CODE

The built-in Unsigned Integer types available for encodings.

UInt8 for bitwidths <= 8

UInt16 for bitwidths > 8

see also FLOAT

FloatsForML.FLOAT_TYPES — Constant

FLOAT

The built-in floating-point types available for valuations.

Float32 for bitwidths <= 8

Float64 for bitwidths > 8

see also CODE

FloatsForML.FLOAT — Type

FLOAT

The built-in floating-point types available for valuations.

Float32 for bitwidths <= 8

Float64 for bitwidths > 8

see also CODE

FloatsForML.MLFloats — Method

MLFloats

examples

``` UF42 = MLFloats( 4, 2, IsUnsigned, IsFinite) UE64 = MLFloats( 6, 4, IsUnsigned, IsExtended) SE84 = MLFloats( 8, 4, IsSigned, IsExtended) SE1512 = MLFloats(15, 12, IsSigned, IsExtended)

seBinary84_encodings = codes(SE84)
seBinary84_valuation = floats(SE84)

````

FloatsForML.index_to_offset — Method

index_to_offset(x)

convert the Julia index x into a P3109 encoding value as a UInt16

>> (0x0100) ↦ 0x00ff

in 0-based languages, this is a do nothing operation

FloatsForML.offset_to_index — Method

offset_to_index(x)

convert the P3109 encoding x into a Julia index as an UInt16

>> (0xff) ↦ 0x0100

in 0-based languages, this is a do nothing operation

FloatsForML.typeforcode — Function

typeforcode(bitwidth)

The bitstype to be used for encoding values of bitwidth

It is an unchecked error to set bitwidth outside BitsMin..BitsMax

FloatsForML.typeforfloat — Function

typeforfloat(bitwidth)

The bitstype to be used for storing values of bitwidth

It is an unchecked error to set bitwidth outside BitsMin..BitsMax