Reference

Contents

Reference

Contents

Index

Index

AIFloats.CODE_TYPES

AIFloats.FLOAT_TYPES

AIFloats.AbstractAIFloat

AIFloats.AbstractSigned

AIFloats.AbstractUnsigned

AIFloats.AkoSignedExtended

AIFloats.AkoSignedFinite

AIFloats.AkoUnsignedExtended

AIFloats.AkoUnsignedFinite

AIFloats.CODE

AIFloats.FLOAT

AIFloats.AIFloat

AIFloats.code_to_index

AIFloats.code_to_value

AIFloats.codes

AIFloats.floats

AIFloats.index_to_code

AIFloats.index_to_value

AIFloats.safe_max

AIFloats.typeforcode

AIFloats.typeforfloat

AIFloats.validate_code

AIFloats.validate_index

AIFloats.value_to_code

AIFloats.value_to_index

AIFloats.x_or_T

AIFloats.CODE — Type

CODE

The built-in Unsigned Integer types available for encodings.

UInt8 for bitwidths <= 8

UInt16 for bitwidths > 8

see also FLOAT

AIFloats.CODE_TYPES — Constant

CODE

The built-in Unsigned Integer types available for encodings.

UInt8 for bitwidths <= 8

UInt16 for bitwidths > 8

see also FLOAT

AIFloats.FLOAT — Type

FLOAT

The built-in floating-point types available for valuations.

Float32 for bitwidths <= 7

Float64 for bitwidths > 8

Float128 for bitwidths > 11

see also CODE

AIFloats.FLOAT_TYPES — Constant

FLOAT

The built-in floating-point types available for valuations.

Float32 for bitwidths <= 7

Float64 for bitwidths > 8

Float128 for bitwidths > 11

see also CODE

AIFloats.AbstractAIFloat — Type

AbstractAIFloat{bitwidth, precision}

supertype: AbstractFloat

see AbstractSigned, AbstractUnsigned

AIFloats.AbstractSigned — Type

AbstractSigned{bitwidth, precision}

supertype: AbstractAIFloat

see AbstractAIFloat, AbstractUnsigned

AIFloats.AbstractUnsigned — Type

AbstractUnsigned{bitwidth, precision}

supertype: AbstractAIFloat

see AbstractAIFloat, AbstractSigned

AIFloats.AkoSignedExtended — Type

AkoSignedExtended{bitwidth, precision}

supertype: AbstractSigned

see AIFloat, AkoSignedFinite, AbstractSigned,

AIFloats.AkoSignedFinite — Type

AkoSignedFinite{bitwidth, precision}

supertype: AbstractSigned

see AIFloat, AkoSignedExtended, AbstractSigned,

AIFloats.AkoUnsignedExtended — Type

AkoUnsignedExtended{bitwidth, precision}

supertype: AbstractUnsigned

see AIFloat, AkoUnsignedFinite, AbstractUnsigned

AIFloats.AkoUnsignedFinite — Type

AkoUnsignedFinite{bitwidth, precision}

supertype: AbstractUnsigned

see AIFloat, AkoUnsignedExtended, AbstractUnsigned,

AIFloats.AIFloat — Function

AIFloat(<:AbstractAIFloat)

AIFloat(bitwidth, precision, `<signedness>`, `<domain`)

bitwidth is the total number of bits in the floating-point representation

precision is the number of bits in the significand (implicit bit + fractional bits)

<signedness> is one of {:signed, :unsigned}

<domain> is one of {:finite, :extended}

generates a corresponding concrete AI floating-point type

see floats, codes

AIFloats.code_to_index — Function

code_to_index(n, code)

convert the P3019 encoding to a Julia index

(64, 0x1e) ↦ 31

(32, 0x1d) ↦ 30

AIFloats.code_to_value — Function

code_to_value(values, code)

- values[code+1] i f   0 <= code < length(values)
                 else NaN

AIFloats.codes — Function

codes(x_or_T)

the canonical encoding sequence for the given type.

see AIFloat, floats, x_or_T

AIFloats.floats — Function

floats(x_or_T)

the canonical value sequence for the given type.

see AIFloat, codes, x_or_T

AIFloats.index_to_code — Function

index_to_code(n, index)

convert the Julia index (1-based) into a P3109 encoding (0-based)

(8, 31) ↦ 0x1e

(9, 30) ↦ 0x1d

AIFloats.index_to_value — Function

index_to_value(values, idx)

- values[idx] if   0 < idx <= length(values)
              else NaN

AIFloats.safe_max — Function

safe_max(x)

allows fma via emulation

AIFloats.typeforcode — Function

typeforcode(bitwidth)

The bitstype to be used for encoding values of bitwidth

It is an unchecked error to set bitwidth outside BitsMin..BitsMax

AIFloats.typeforfloat — Function

typeforfloat(bitwidth)

The bitstype to be used for storing values of bitwidth

It is an unchecked error to set bitwidth outside BitsMin..BitsMax

AIFloats.validate_code — Function

validate_code(n::Integer, code::Unsigned)

code or nothing

AIFloats.validate_index — Function

validate_index(n::Integer, code::Unsigned)

code or nothing

AIFloats.value_to_code — Function

value_to_code

AIFloats.value_to_index — Function

value_to_index

AIFloats.x_or_T — Function

x_or_T

func(x_or_T) means func works with abstract and concrete arguments

the concrete x::T

the abstract T::Type